projects/Survey-Chat-Bot/requirements.txt
# Survey Chat Bot - Core dependencies
streamlit>=1.55.0
google-generativeai>=0.8.6
python-dotenv>=1.0.1
